deleted a lot of work
following the logical structure (i.e. this document needs to be comprehensible for readers) i needed to reconsider the structuring of this section first and  am now able to list problems in an easier to understand way.

building this chapter is difficult, if you're too familiar with the topic. i had sadly been trying to work from "difficult" to "easy", which does not work in this circumstance, as it does not provide logical structure for a thesis... i mean i could technically reverse the ordering afterwards, but that is... an issue for the future